dashboard/dashboard_requirements.txt
python-dotenv
redshift_connector
matplotlib
streamlit
pandas